--- src/pages/forgot-password/forgot-password.scss
@@ -1,32 +1,0 @@
-header {
-  width: 100%;
-  padding-top: 124px;
-  text-align: center;
-
-  h1 {
-    line-height: 48px;
-    font-size: 36px;
-    font-weight: 300;
-    color: #554d56;
-  }
-
-  h2 {
-    line-height: 24px;
-    font-size: 18px;
-    color: #ee6723;
-  }
-}
-
-form.forgot-password {
-  position: relative;
-  display: block;
-  max-width: 392px;
-  width: 100%;
-  margin: 12px auto 0;
-  button {
-    height: 60px;
-    background-color: #fdbd39;
-    color: #fff;
-    margin: 8px 16px 12px;
-  }
-}
