+++ src/pages/index.js
@@ -1,3 +1,3 @@
-export { default as Dashboard } from './dashboard'
-export { default as Signin } from './signin'
-export { default as Signup } from './signup'
+export { default as Dashboard } from './dashboard';
+export { default as Signin } from './signin/signinForm';
+export { default as Signup } from './signup/signupForm';
