--- src/pages/signup/signup.scss
@@ -1,32 +1,0 @@
-.termsOfService {
-  display: flex;
-  align-items: center;
-  line-height: 19px;
-  font-size: 14px;
-  color: #979197;
-  cursor: pointer;
-  padding: 27px 16px 4px;
-
-  input {
-    display: block;
-    flex-grow: 0;
-    opacity: 0;
-    height: 0;
-  }
-}
-
-.checkbox {
-  width: 16px;
-  height: 16px;
-  background-size: contain;
-  background-position: center;
-  background-repeat: no-repeat;
-  margin-right: 6px;
-}
-
-.checkbox-on {
-  background-image: url('../../assets/icons/icCheckOn.svg');
-}
-.checkbox-off {
-  background-image: url('../../assets/icons/icCheckOff.svg');
-}
